requirements.txt
Django==1.7.4
django-filter==0.9.2
djangorestframework==3.0.4
jsonfield==1.0.2